feat: Commit 구현 대략 완료, 추후 버그 탐색 필요

+++ src/components/PersonalReading/sideComponents/HighlightList/HighlightList.js
@@ -15,22 +15,38 @@ function HighlightList(props) {
     let [highlightData, setHighlightData] = useState([]);
     
     useEffect(() => {
-        async function getHighlightData() {
-            await axios.get(`http://43.200.26.215:3000/highlights/pdfs/${props.pdfIdx}/pages/${props.currentPageNumber}`)
-                .then((response) => {
-                    // let result = Array();
-                    
-                    // for (let i = 0; i < response.data.result.length; i++) {
-                    //     if (response.data.result[i].active === 1) {
-                    //         result.push(response.data.result[i])
-                    //     }
-                    // }
-                    setHighlightData(response.data.result);
-                });
+        if (props.commitIdx === -1) {
+            async function getHighlightData() {
+                await axios.get(`http://43.200.26.215:3000/highlights/pdfs/${props.pdfIdx}/pages/${props.currentPageNumber}`)
+                    .then((response) => {
+                        // let result = Array();
+                        
+                        // for (let i = 0; i < response.data.result.length; i++) {
+                        //     if (response.data.result[i].active === 1) {
+                        //         result.push(response.data.result[i])
+                        //     }
+                        // }
+                        setHighlightData(response.data.result);
+                    });
+            }
+            
+            getHighlightData();
         }
-
-        getHighlightData();
-    }, [props.currentPageNumber, props.updateHighlightList]);
+        
+        else {
+            console.log(props.pdfIdx);
+            console.log(props.currentPageNumber);
+            console.log(props.commitIdx);
+            async function getHighlightData() {
+                await axios.get(`http://43.200.26.215:3000/highlights/pdfs/${props.pdfIdx}/pages/${props.currentPageNumber}/commitIdx/${props.commitIdx}`)
+                    .then((response) => {
+                        setHighlightData(response.data.result);
+                    })
+            }
+            
+            getHighlightData();
+        }
+    }, [props.currentPageNumber, props.updateHighlightList, props.commitIdx]);
 
     return (
         <>
@@ -43,6 +59,8 @@ function HighlightList(props) {
             <aside className={styles.wrap}>
                 <div className={styles.container}>
                     <HighlightCards
+                        commitIdx={props.commitIdx}
+                        pdfIdx={props.pdfIdx}
                         highlightData={highlightData}
                         setHighlightData={setHighlightData}
                         style={{ overflow: "scroll" }}
@@ -61,34 +79,52 @@ function HighlightCards(props) {
         event.dataTransfer.setData("text/plain", event.target.innerHTML);
     }
 
-    async function deleteHighlight(highlightIdx, setHighlightData, updateHighlightList, setUpdateHighlightList, currentPageNumber) {
-        await axios.delete(`http://43.200.26.215:3000/highlights/${highlightIdx}`)
-            .then((response) => {
-                console.log("highlight delete response:", response);
-                
-            })
-            .catch((error) => {
-                console.log("highlight delete error:", error);
-            })
-            .then(async () => {
-                await axios.get(`http://43.200.26.215:3000/highlights/pdfs/${74}/pages/${currentPageNumber}`)
-                    .then((response) => {
-                        // let result = Array();
-                        // for (let i = 0; i < response.data.result.length; i++) {
-                        //     if (response.data.result[i].active === 1) {
-                        //         result.push(response.data.result[i]);
-                        //     }
-                        // }
-                        
-                        setHighlightData(response.data.result);
-                        setUpdateHighlightList(!updateHighlightList);
-                        const selectedHighlight = document.getElementsByClassName('highlight' + highlightIdx);
-                        
-                        for (let i = 0; i < selectedHighlight.length; i++) {
-                            selectedHighlight[i].classList.remove('highlighted', `${highlightIdx}`);
-                        }
-                    })
-            })
+    async function deleteHighlight(commitIdx, highlightIdx, setHighlightData, updateHighlightList, setUpdateHighlightList, currentPageNumber) {
+        
+        if (commitIdx === -1) {
+            
+            await axios.delete(`http://43.200.26.215:3000/highlights/${highlightIdx}`)
+                .then((response) => {
+                    console.log("highlight delete response:", response);
+                })
+                .catch((error) => {
+                    console.log("highlight delete error:", error);
+                })
+                .then(async () => {
+                    await axios.get(`http://43.200.26.215:3000/highlights/pdfs/${props.pdfIdx}/pages/${currentPageNumber}`)
+                        .then((response) => {
+                            setHighlightData(response.data.result);
+                            setUpdateHighlightList(!updateHighlightList);
+                            const selectedHighlight = document.getElementsByClassName('highlight' + highlightIdx);
+                            
+                            for (let i = 0; i < selectedHighlight.length; i++) {
+                                selectedHighlight[i].classList.remove('highlighted', `${highlightIdx}`);
+                            }
+                        })
+                })
+        } else {
+            
+            // await axios.delete(`http://43.200.26.215:3000/highlights/${highlightIdx}`)
+            //     .then((response) => {
+            //         console.log("highlight delete response:", response);
+            //     })
+            //     .catch((error) => {
+            //         console.log("highlight delete error:", error);
+            //     })
+            //     .then(async () => {
+            //         await axios.get(`http://43.200.26.215:3000/highlights/pdfs/${pdfIdx}/pages/${currentPageNumber}/commitIdx/${commitIdx}`)
+            //             .then((response) => {
+            //                 setHighlightData(response.data.result);
+            //                 setUpdateHighlightList(!updateHighlightList);
+            //                 const selectedHighlight = document.getElementsByClassName('highlight' + highlightIdx);
+                            
+            //                 for (let i = 0; i < selectedHighlight.length; i++) {
+            //                     selectedHighlight[i].classList.remove('highlighted', `${highlightIdx}`);
+            //                 }
+            //             })
+            //     })
+        }
+        
     }
 
     return (
@@ -102,7 +138,7 @@ function HighlightCards(props) {
                                 avatar={<Avatar sx={{ bgcolor: "#4DABB3", width: 10, height: 10 }} aria-label="recipe">{""}</Avatar>}
                                 title={<p style={{ color: "#DDDDDD" }}></p>}
                                 action={
-                                    <IconButton onClick={() => { deleteHighlight( element.highlightIdx, props.setHighlightData, props.updateHighlightList, props.setUpdateHighlightList, props.currentPageNumber ); }}>
+                                    <IconButton onClick={() => { deleteHighlight( props.commitIdx, element.highlightIdx, props.setHighlightData, props.updateHighlightList, props.setUpdateHighlightList, props.currentPageNumber ); }}>
                                         <ClearIcon fontSize="small"></ClearIcon>
                                     </IconButton>
                                 }
